fixed import paths in scripts and tests folders
