+++ scripts/build-newsroom-videos.ts
@@ -5,7 +5,7 @@ import { dirname, resolve } from 'path';
 import process from 'process';
 import { fileURLToPath } from 'url';
 
-import { logger } from './utils/logger';
+import { logger } from './utils';
 
 const currentFilePath = fileURLToPath(import.meta.url);
 const currentDirPath = dirname(currentFilePath);
