+++ scripts/compose.ts
@@ -7,7 +7,7 @@ import fs from 'fs';
 import inquirer from 'inquirer';
 import moment from 'moment';
 
-import { logger } from './utils/logger';
+import { logger } from './utils';
 
 /**
  * Type definition for the answers from the compose prompt.
