+++ scripts/markdown/check-edit-links.ts
@@ -5,7 +5,7 @@ import { fileURLToPath } from 'url';
 
 import editUrls from '../../config/edit-page-config.json';
 import { pause } from '../utils';
-import { logger } from '../utils/logger';
+import { logger } from '../utils';
 
 const ignoreFiles = [
   'reference/specification/v2.x.md',
