+++ scripts/markdown/check-markdown.ts
@@ -3,7 +3,7 @@ import matter from 'gray-matter';
 import path, { dirname } from 'path';
 import { fileURLToPath } from 'url';
 
-import { logger } from '../utils/logger';
+import { logger } from '../utils';
 
 const currentFilePath = fileURLToPath(import.meta.url);
 const currentDirPath = dirname(currentFilePath);
