+++ scripts/tools/extract-tools-github.ts
@@ -5,7 +5,7 @@ import dotenv from 'dotenv';
 import type { ToolsData } from '@/types/scripts/tools';
 
 import { pause } from '../utils';
-import { logger } from '../utils/logger';
+import { logger } from '../utils';
 
 dotenv.config();
