--- scripts/utils/logger.ts
@@ -1,18 +1,0 @@
-import winston from 'winston';
-
-const { combine, timestamp, printf, colorize, align } = winston.format;
-
-const logger = winston.createLogger({
-  level: process.env.LOG_LEVEL || 'info',
-  format: combine(
-    colorize({ level: true }),
-    timestamp({
-      format: 'YYYY-MM-DD hh:mm:ss.SSS A'
-    }),
-    align(),
-    printf((info) => `[${info.timestamp}] ${info.level}: ${info.message}`)
-  ),
-  transports: [new winston.transports.Console()]
-});
-
-export { logger };
