+++ tests/tools/combine-tools.test.js
@@ -20,7 +20,7 @@ const {
   circularTool,
   finalToolWithMissingData
 } = require('../fixtures/combineToolsData');
-const { logger } = require('../../scripts/utils/logger.ts');
+import { logger } from '../../scripts/utils.ts';
 
 jest.mock('../../scripts/utils/logger', () => ({
   logger: { error: jest.fn() }
